Add pkg dir and move funcs out of main
